updated dataset

+++ data/deam.py
@@ -1,5 +1,7 @@
 from os import path
 
+import numpy as np
+
 import torch
 import torchaudio
 
@@ -37,10 +39,10 @@ class DeamDataset(BaseAudioDataset):
             return label
         if self.label_type == "static":
             ret_cols = ['static_valence_mean', 'static_valence_std', 'static_arousal_mean', 'static_arousal_std']
-            return torch.tensor(item[ret_cols])
+            return torch.tensor(item[ret_cols].astype(np.float32))
         if self.label_type == "dynamic":
             ret_cols = ['dynamic_valence_mean', 'dynamic_valence_std', 'dynamic_arousal_mean', 'dynamic_arousal_std']
-            return torch.tensor(item[ret_cols])
+            return torch.tensor(item[ret_cols].astype(np.float32))
         raise NameError
 
     def get_audio(self, index):
